+++ Core/Factories/VehiclePresenterFactory.cs
@@ -1,4 +1,4 @@
 namespace Core.Factories
 {
-  public delegate Presenters.VehiclePresenter VehiclePresenterFactory(ushort trainAddress);
+  public delegate Presenters.VehiclePresenter VehiclePresenterFactory(int trainAddress);
 }
